Empty user ID or replica Set ID for delete Rs.

+++ modules/repl-server/src/main/java/edu/emory/bmi/datarepl/servlets/DeleteRsServlet.java
@@ -26,25 +26,32 @@ public class DeleteRsServlet extends HttpServlet {
         response.setContentType("text/html");
         PrintWriter out = response.getWriter();
 
-        long replicaSetID = Long.parseLong(request.getParameter("replicaSetID"));
         String userId = request.getParameter("userID");
+        if (request.getParameter("replicaSetID") != null && !request.getParameter("replicaSetID").equals("") &&
+                !userId.trim().equals("")) {
+            long replicaSetID = Long.parseLong(request.getParameter("replicaSetID"));
 
-        TciaReplicaSetInterface tciaReplicaSetInterface = (TciaReplicaSetInterface) TciaReplicaSetInterface.getInfiniCore();
+            TciaReplicaSetInterface tciaReplicaSetInterface = (TciaReplicaSetInterface) TciaReplicaSetInterface.getInfiniCore();
 
-        logger.info("Deleting the replica set of the user..");
-        boolean deleted = tciaReplicaSetInterface.deleteReplicaSet(userId, replicaSetID);
+            logger.info("Deleting the replica set of the user..");
+            boolean deleted = tciaReplicaSetInterface.deleteReplicaSet(userId, replicaSetID);
 
-        if (deleted) {
-            Long[] replicaSets = tciaReplicaSetInterface.getUserReplicaSets(userId);
+            if (deleted) {
+                Long[] replicaSets = tciaReplicaSetInterface.getUserReplicaSets(userId);
 
-            String output = UIGenerator.returnReplicaSetOutput(replicaSets);
+                String output = UIGenerator.returnReplicaSetOutput(replicaSets);
 
-            logger.info("Listing the Replica Sets of the User");
-            out.println(output);
+                logger.info("Listing the Replica Sets of the User");
+                out.println(output);
+            } else {
+                out.println("<HTML>    <BODY>\n");
+                out.println("Error in Deleting the replicaSet with ReplicaSet ID: " + replicaSetID);
+                out.println("</body></html>");
+            }
         } else {
-            out.println("<HTML>    <BODY>\n");
-            out.println("Error in Deleting the replicaSet with ReplicaSet ID: " + replicaSetID);
-            out.println("</body></html>");
+            String output = "Empty values provided for the user ID or the replica Set ID";
+            logger.error(output);
+            out.println(output);
         }
     }
 }
